fix the fxxking bugs 1. trap.c p->killed = - 1 -> 1 2. uint16 -> int kcnt 3. mappage parameters va -> PGROUNDDOWN(va) fuck off

+++ kernel/kalloc.c
@@ -29,7 +29,7 @@ struct {
 struct spinlock kcnt_lock;
 #define PA2KCNT_IDX(p) (((p) - KERNBASE) / PGSIZE)
 #define KCNT_MAX_IDX PA2KCNT_IDX(PHYSTOP)
-uint16 kcnt[KCNT_MAX_IDX] = {0};
+int kcnt[KCNT_MAX_IDX] = {0};
 #define PA2KCNT(p) kcnt[PA2KCNT_IDX((uint64)(p))]
 
 void
